linkedin-zip-challenge/pyproject.toml
[project]
name = "linkedin-zip-challenge"
version = "0.1.0"
description = "A solver for the LinkedIn Zip Challenge."
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.119.0",
    "gradio>=5.49.1",
    "gymnasium>=1.2.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "numpy>=1.26.4,<2.0",
    "ortools>=9.14.6206",
    "pillow>=11.3.0",
    "pydantic-settings>=2.11.0",
    "pytest>=8.4.2",
    "pytest-html>=4.1.1",
    "pytest-reportlog>=0.4.0",
    "python-multipart>=0.0.20",
    "sortedcontainers>=2.4.0",
    "stable-baselines3>=2.7.0",
    "tensorboard>=2.20.0",
    "torch==2.4.1",
    "torchaudio==2.4.1",
    "torchvision==0.19.1",
    "tqdm>=4.67.1",
    "uvicorn[standard]>=0.37.0",
]

[tool.isort]
profile = "black"


[tool.uv.pip]
extra-index-url = ["https://download.pytorch.org/whl/cu121"]
index-strategy = "unsafe-best-match"

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu121"

[[tool.uv.index]]
url = "https://pypi.org/simple"
